renamed the namespace for the retry connection stuff to be more clear, fixing MD issues
