+++ src/EntityManager/EntityManagerFactory.php
@@ -11,8 +11,8 @@ use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Tools;
 use EdmondsCommerce\DoctrineStaticMeta\ConfigInterface;
 use EdmondsCommerce\DoctrineStaticMeta\Entity\Factory\EntityFactoryInterface;
-use EdmondsCommerce\DoctrineStaticMeta\EntityManager\Connection\RetryConnection;
-use EdmondsCommerce\DoctrineStaticMeta\EntityManager\Connection\ShouldConnectionByRetried;
+use EdmondsCommerce\DoctrineStaticMeta\EntityManager\RetryConnection\RetryConnection;
+use EdmondsCommerce\DoctrineStaticMeta\EntityManager\RetryConnection\ShouldConnectionByRetried;
 use EdmondsCommerce\DoctrineStaticMeta\EntityManager\Decorator\EntityFactoryManagerDecorator;
 use EdmondsCommerce\DoctrineStaticMeta\Exception\DoctrineStaticMetaException;
 use EdmondsCommerce\DoctrineStaticMeta\MappingHelper;
